Assignment1

+++ include/cs474.h
@@ -10,8 +10,8 @@ namespace cs474 {
 	std::optional<std::vector<uint8_t>> scale(const std::vector<uint8_t>& subsampleData, int factor, int width, int height);
 	std::optional<std::vector<uint8_t>> quantize(const std::vector<uint8_t>& data, uint8_t levels, int width, int height);
 	std::vector<float> hist(const std::vector<uint8_t>& data);
-	std::vector<size_t> cdf(const std::vector<uint8_t>& data);
-	std::vector<uint8_t> norm(const std::vector<uint8_t>& data, int total_pixels);
+	std::vector<float> cdf(const std::vector<float>& data);
+	std::vector<uint8_t> norm(const std::vector<float>& data, int total_pixels);
 	std::vector<uint8_t> hist_eq(const std::vector<uint8_t>& data, int width, int height);
 	std::optional<std::vector<uint8_t>> hist_spec(const std::vector<uint8_t>& srcData, const std::vector<uint8_t>& specData, int width, int height);
-}
+} // namespace cs474
